Organized header files and fixed bug in RC_ChangeMethodCodeAttribute
Many changes to the place where funcions were in the header files. Now every
exported funcion stays in reclass.h
Fixed a bug on RC_ChangeMethodCodeAttribute where the classfile would not
be recognized anymowhere after its code attribute was changed. This is due
the fact that the LineNumberTable attribute was pointing to a part of the
code attribute that did not exist anymore.